move test_case module
