+++ tests/units/anta_tests/routing/test_generic.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 from typing import Any
 
 from anta.tests.routing.generic import VerifyBFD, VerifyRoutingProtocolModel, VerifyRoutingTableSize
-from tests.units.anta_tests.test_case import test
+from tests.lib.test_case import test
 
 DATA: list[dict[str, Any]] = [
     {
