+++ tests/units/anta_tests/routing/test_ospf.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 from typing import Any
 
 from anta.tests.routing.ospf import VerifyOSPFNeighborCount, VerifyOSPFNeighborState
-from tests.units.anta_tests.test_case import test
+from tests.lib.test_case import test
 
 DATA: list[dict[str, Any]] = [
     {
