+++ tests/units/anta_tests/test_configuration.py
@@ -13,7 +13,7 @@ from __future__ import annotations
 from typing import Any
 
 from anta.tests.configuration import VerifyRunningConfigDiffs, VerifyZeroTouch
-from tests.units.anta_tests.test_case import test
+from tests.lib.test_case import test
 
 DATA: list[dict[str, Any]] = [
     {
