+++ tests/units/anta_tests/test_connectivity.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 from typing import Any
 
 from anta.tests.connectivity import VerifyReachability
-from tests.units.anta_tests.test_case import test
+from tests.lib.test_case import test
 
 DATA: list[dict[str, Any]] = [
     {
