+++ tests/units/anta_tests/test_stp.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 from typing import Any
 
 from anta.tests.stp import VerifySTPBlockedPorts, VerifySTPCounters, VerifySTPForwardingPorts, VerifySTPMode, VerifySTPRootPriority
-from tests.units.anta_tests.test_case import test
+from tests.lib.test_case import test
 
 DATA: list[dict[str, Any]] = [
     {
