+++ tests/units/anta_tests/test_system.py
@@ -17,7 +17,7 @@ from anta.tests.system import (
     VerifySyslog,
     VerifyUptime,
 )
-from tests.units.anta_tests.test_case import test
+from tests.lib.test_case import test
 
 DATA: list[dict[str, Any]] = [
     {
